refactor: add new lib folder to src folder
