+++ instagram-app/src/lib/actionCreators.js
@@ -1,0 +1,1 @@
+// TODO: add action creators here
